chore: rename model_features.js -> models.js
